Added company registration and enabling. Added password policy component. Started draft implementation for the login

+++ src/main/java/mx/nopaloverflow/petservices/dtos/AddressDto.java
@@ -7,8 +7,8 @@ public class AddressDto {
     private String stateCode;
     private String countryIsoCode;
     private String zipCode;
-    private Long latitude;
-    private Long longitude;
+    private Double latitude;
+    private Double longitude;
 
     public String getStreet() {
         return street;
@@ -50,19 +50,19 @@ public class AddressDto {
         this.zipCode = zipCode;
     }
 
-    public Long getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Long latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public Long getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Long longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 }
